Export InferenceProviderMappingEntry from types.ts
